refactor(raijin): split initializer responsibilities
