+++ yarn/raijin/src/initializer.ts
@@ -1,0 +1,32 @@
+import type { FetchLike }         from './runtime-download.js'
+import type { YarnCommandRunner } from './yarn-command.js'
+
+import { installRaijinRuntime }   from './runtime-installer.js'
+import { runYarnCommand }         from './yarn-command.js'
+
+const CODE_RUNTIME_PACKAGE = '@atls/code-runtime@latest'
+
+export interface RunRaijinInitializerOptions {
+  argv?: Array<string>
+  cwd?: string
+  fetchImpl?: FetchLike
+  runYarnCommand?: YarnCommandRunner
+}
+
+export const runRaijinInitializer = async ({
+  argv = [],
+  cwd = process.cwd(),
+  fetchImpl = fetch,
+  runYarnCommand: runCommand = runYarnCommand,
+}: RunRaijinInitializerOptions = {}): Promise<void> => {
+  const command = argv[0]
+
+  if (argv.length > 1 || (command && command !== 'init')) {
+    throw new Error('Usage: yarn init @atls/raijin or yarn dlx @atls/raijin init')
+  }
+
+  await installRaijinRuntime({ cwd, fetchImpl })
+  await runCommand(['add', '-D', CODE_RUNTIME_PACKAGE], cwd)
+  await runCommand(['generate', 'project'], cwd)
+  await runCommand(['tools', 'sync'], cwd)
+}
